# public/docs/openapi.yaml
openapi: 3.1.0
info:
  title: 'Investment API'
  description: 'JSON API for creating, listing, viewing, and withdrawing investments.'
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000'
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'Send a Sanctum Bearer token in the `Authorization` header. **Admin** may access every investment; **Owner** may access only their own. Endpoints that do not require a token are marked as unauthenticated.'
security:
  -
    default: []
paths:
...
  /api/investments:
    get:
      summary: 'List investments'
      operationId: listInvestments
      description: "Paginated investments visible to the current user. Owners see only their own\nrecords. Admins see every owner's investments. Query user_id is ignored."
      parameters:
        -
          in: query
          name: page
          description: 'Page number.'
          required: false
          schema:
            type: integer
            description: 'Page number.'
            examples:
              - 1
        -
          in: query
          name: per_page
          description: 'Items per page (1-100, default 15).'
          required: false
          schema:
            type: integer
            description: 'Items per page (1-100, default 15).'
            examples:
              - 15
      responses:
        200:
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    examples:
                      - []
                  links:
                    type: object
                    properties: {}
                  meta:
                    type: object
                    properties:
                      current_page:
                        type: integer
                        examples:
                          - 1
                      per_page:
                        type: integer
                        examples:
                          - 15
                      total:
                        type: integer
                        examples:
                          - 0
                examples:
                  -
                    data: []
                    links: {}
                    meta:
                      current_page: 1
                      per_page: 15
                      total: 0
        401:
          description: unauthenticated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    examples:
                      - Unauthenticated.
                examples:
                  -
                    message: Unauthenticated.
      tags:
        - Investments
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                page:
                  type: integer
                  description: 'Must be at least 1.'
                  examples:
                    - 16
                per_page:
                  type: integer
                  description: 'Must be at least 1. Must not be greater than 100.'
                  examples:
                    - 22
    post:
      summary: 'Create investment'
      operationId: createInvestment
      description: "Registers an investment owned by the authenticated user. Extra fields such as\nuser_id, status, expected_balance, tax, or withdrawn_on are ignored."
      parameters: []
      responses:
        201:
          description: created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        examples:
                          - 1
                      owner:
                        type: object
                        properties:
                          id:
                            type: integer
                            examples:
                              - 2
                          name:
                            type: string
                            examples:
                              - Owner
                          email:
                            type: string
                            examples:
                              - owner@example.com
                      amount:
                        type: string
                        examples:
                          - '1000.00'
                      created_on:
                        type: string
                        examples:
                          - '2025-01-15'
                      status:
                        type: string
                        examples:
                          - active
                      withdrawn_on:
                        type:
                          - string
                          - 'null'
                        examples:
                          - null
                      expected_balance:
                        type: string
                        examples:
                          - '1000.00'
                      gain:
                        type: string
                        examples:
                          - '0.00'
                      tax:
                        type: string
                        examples:
                          - '0.00'
                      net:
                        type: string
                        examples:
                          - '1000.00'
                      rate:
                        type: string
                        examples:
                          - '0.225'
                      complete_months:
                        type: integer
                        examples:
                          - 0
                examples:
                  -
                    data:
                      id: 1
                      owner:
                        id: 2
                        name: Owner
                        email: owner@example.com
                      amount: '1000.00'
                      created_on: '2025-01-15'
                      status: active
                      withdrawn_on: null
                      expected_balance: '1000.00'
                      gain: '0.00'
                      tax: '0.00'
                      net: '1000.00'
                      rate: '0.225'
                      complete_months: 0
        401:
          description: unauthenticated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    examples:
                      - Unauthenticated.
                examples:
                  -
                    message: Unauthenticated.
        422:
          description: invalid
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    examples:
                      - 'The amount field format is invalid.'
                  errors:
                    type: object
                    properties:
                      amount:
                        type: array
                        items:
                          type: string
                        examples:
                          -
                            - 'The amount field format is invalid.'
                examples:
                  -
                    message: 'The amount field format is invalid.'
                    errors:
                      amount:
                        - 'The amount field format is invalid.'
      tags:
        - Investments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: string
                  description: 'Amount with two decimal places, greater than 0.00.'
                  examples:
                    - '1000.00'
                created_on:
                  type: date
                  description: 'Creation date (today or past).'
                  examples:
                    - '2025-01-15'
              required:
                - amount
                - created_on
...